Document the IM model; split the im.rs file into sub-modules
